refactor: Move InvalidTokenCache to package org.knora.webapi.slice.infrastructure
